end_to_end/TPU/resnet-research-JAX-Distributed-Shampoo-tpu-v4-256/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
attrs==21.2.0
build==0.6.0.post1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
chex==0.0.8
cloudpickle==1.6.0
clu==0.0.6
contextlib2==21.6.0
cvxpy==1.1.15
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
dm-tree==0.1.6
ecos==2.0.7.post1
flatbuffers==1.12
flax==0.3.4
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphsurgeon==0.4.5
grpcio==1.39.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
jax==0.2.19
jaxlib==0.1.70
keras-nightly==2.6.0.dev2021062500
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==11.1.0
libtpu-nightly==0.1.dev20210809
Markdown==3.3.4
matplotlib==3.4.3
ml-collections==0.1.0
msgpack==1.0.2
networkx==2.6.2
numpy==1.19.5
nvidia-pyindex==1.0.9
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.0.9
osqp==0.6.2.post0
packaging==21.0
pandas==1.3.2
pep517==0.11.0
Pillow==8.3.1
plumber-analysis-mkuchnik @ file:///home/mkuchnik_andrew_cmu_edu/tfplumber_app/example_pipelines/TF2_Simple/plumber_analysis/dist/plumber_analysis_mkuchnik-0.0.1-py3-none-any.whl
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
qdldl==0.1.5.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
scs==2.1.4
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow @ file:///home/mkuchnik_andrew_cmu_edu/tensorflow-2.7.0-cp37-cp37m-linux_x86_64.whl
tensorflow-datasets==4.4.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
tfa-nightly==0.15.0.dev20210905153013
tfp-nightly==0.14.0.dev20210909
tomli==1.2.1
toolz==0.11.1
tqdm==4.62.2
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
